+++ core/common/role.go
@@ -1,16 +1,22 @@
-package permissions
+package common
 
 import (
 	"net/http"
 
-	corecommon "github.com/arran4/goa4web/core/common"
-	hcommon "github.com/arran4/goa4web/handlers/common"
 	dbpkg "github.com/arran4/goa4web/internal/db"
+)
+
+// Context keys used by Allowed when reading from the request context.
+const (
+	// KeyCoreData provides access to CoreData.
+	KeyCoreData ContextValues = "coreData"
+	// KeyQueries holds the db.Queries pointer.
+	KeyQueries ContextValues = "queries"
 )
 
 // Allowed checks if the request context provides one of the given roles.
 func Allowed(r *http.Request, roles ...string) bool {
-	cd, ok := r.Context().Value(hcommon.KeyCoreData).(*corecommon.CoreData)
+	cd, ok := r.Context().Value(KeyCoreData).(*CoreData)
 	if ok && cd != nil {
 		for _, lvl := range roles {
 			if cd.HasRole(lvl) {
@@ -20,7 +26,7 @@ func Allowed(r *http.Request, roles ...string) bool {
 		return false
 	}
 
-	queries, qok := r.Context().Value(hcommon.KeyQueries).(*dbpkg.Queries)
+	queries, qok := r.Context().Value(KeyQueries).(*dbpkg.Queries)
 	if !qok {
 		return false
 	}
@@ -35,7 +41,7 @@ func Allowed(r *http.Request, roles ...string) bool {
 	if err != nil || !roleVal.Valid {
 		return false
 	}
-	cd = corecommon.NewCoreData(r.Context(), queries)
+	cd = NewCoreData(r.Context(), queries)
 	cd.SetRole(roleVal.String)
 	for _, lvl := range roles {
 		if cd.HasRole(lvl) {
